Create Code: Average Calculator (Python)
This Python script calculates the average of a list of numbers. It defines a function called "average()" that takes a list as input, adds all the numbers together, and divides the total by the number of elements in the list. If the list is empty, it safely returns "0". The script then uses this function to calculate and display the average of a sample list of grades.